Test methods moved to ActionEvaluatorTest
EvaluateTx tests moved

Code styling
