+++ Libplanet.Tests/Tx/TransactionTest.cs
@@ -2,12 +2,8 @@ using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Diagnostics.CodeAnalysis;
-using System.Linq;
-using System.Numerics;
 using Bencodex.Types;
 using Libplanet.Action;
-using Libplanet.Assets;
-using Libplanet.Blocks;
 using Libplanet.Crypto;
 using Libplanet.Tests.Common.Action;
 using Libplanet.Tx;
@@ -525,198 +521,6 @@ namespace Libplanet.Tests.Tx
                 tx.Actions[1].InnerAction.PlainValue
             );
             Assert.Equal(bytes.Length, tx.BytesLength);
-        }
-
-        [SuppressMessage(
-            "Microsoft.StyleCop.CSharp.ReadabilityRules",
-            "SA1118",
-            Justification = "Long array literals should be multiline.")]
-        [Fact]
-        public void EvaluateActions()
-        {
-            Address[] addresses =
-            {
-                new PrivateKey().ToAddress(),
-                new PrivateKey().ToAddress(),
-                new PrivateKey().ToAddress(),
-            };
-            DumbAction[] actions =
-            {
-                new DumbAction(
-                    targetAddress: addresses[0],
-                    item: "0",
-                    transferFrom: addresses[0],
-                    transferTo: addresses[1],
-                    transferAmount: 5,
-                    recordRandom: true),
-                new DumbAction(
-                    targetAddress: addresses[1],
-                    item: "1",
-                    transferFrom: addresses[2],
-                    transferTo: addresses[1],
-                    transferAmount: 10,
-                    recordRandom: true),
-                new DumbAction(
-                    targetAddress: addresses[0],
-                    item: "2",
-                    transferFrom: addresses[1],
-                    transferTo: addresses[0],
-                    transferAmount: 10,
-                    recordRandom: true),
-                new DumbAction(addresses[2], "R", true, recordRandom: true),
-            };
-            ActionEvaluator<DumbAction> actionEvaluator = new ActionEvaluator<DumbAction>(
-                policyBlockAction: null,
-                stateGetter: ActionEvaluator<DumbAction>.NullStateGetter,
-                balanceGetter: ActionEvaluator<DumbAction>.NullBalanceGetter,
-                trieGetter: null);
-
-            Transaction<DumbAction> tx =
-                Transaction<DumbAction>.Create(0, _fx.PrivateKey1, null, actions);
-            foreach (bool rehearsal in new[] { false, true })
-            {
-                DumbAction.RehearsalRecords.Value =
-                    ImmutableList<(Address, string)>.Empty;
-                var evaluations = actionEvaluator.EvaluateTxGradually(
-                    tx,
-                    default,
-                    1,
-                    new AccountStateDeltaImpl(
-                        address => null,
-                        (a, c) => new FungibleAssetValue(c),
-                        tx.Signer),
-                    addresses[0],
-                    rehearsal: rehearsal).ToImmutableArray();
-
-                Assert.Equal(actions.Length, evaluations.Length);
-                string[][] expectedStates =
-                {
-                    new[] { "0", null, null },
-                    new[] { "0", "1", null },
-                    new[] { "0,2", "1", null },
-                    new[] { "0,2", "1", $"R:{rehearsal}" },
-                };
-                BigInteger[][] expectedBalances =
-                {
-                    new BigInteger[] { -5, 5, 0 },
-                    new BigInteger[] { -5, 15, -10 },
-                    new BigInteger[] { 5, 5, -10 },
-                    new BigInteger[] { 5, 5, -10 },
-                };
-
-                Currency currency = DumbAction.DumbCurrency;
-                IValue[] initStates = new IValue[3];
-                BigInteger[] initBalances = new BigInteger[3];
-                for (int i = 0; i < evaluations.Length; i++)
-                {
-                    ActionEvaluation eval = evaluations[i];
-                    Assert.Equal(actions[i], eval.Action);
-                    Assert.Equal(_fx.Address1, eval.InputContext.Signer);
-                    Assert.Equal(tx.Id, eval.InputContext.TxId);
-                    Assert.Equal(addresses[0], eval.InputContext.Miner);
-                    Assert.Equal(1, eval.InputContext.BlockIndex);
-                    Assert.Equal(rehearsal, eval.InputContext.Rehearsal);
-                    Assert.Equal(
-                        (Integer)eval.OutputStates.GetState(
-                            DumbAction.RandomRecordsAddress),
-                        (Integer)eval.InputContext.Random.Next());
-                    ActionEvaluation prevEval = i > 0 ? evaluations[i - 1] : null;
-                    Assert.Equal(
-                        prevEval is null
-                            ? initStates
-                            : addresses.Select(prevEval.OutputStates.GetState),
-                        addresses.Select(eval.InputContext.PreviousStates.GetState));
-                    Assert.Equal(
-                        expectedStates[i],
-                        addresses.Select(eval.OutputStates.GetState)
-                            .Select(x => x is Text t ? t.Value : null));
-                    Assert.Equal(
-                        prevEval is null
-                            ? initBalances
-                            : addresses.Select(a =>
-                                prevEval.OutputStates.GetBalance(a, currency).RawValue),
-                        addresses.Select(
-                            a => eval.InputContext.PreviousStates
-                                    .GetBalance(a, currency).RawValue));
-                    Assert.Equal(
-                        expectedBalances[i],
-                        addresses.Select(a => eval.OutputStates.GetBalance(a, currency).RawValue));
-                }
-
-                if (rehearsal)
-                {
-                    Assert.Contains(
-                        (addresses[2], "R"),
-                        DumbAction.RehearsalRecords.Value);
-                }
-                else
-                {
-                    Assert.DoesNotContain(
-                        (addresses[2], "R"),
-                        DumbAction.RehearsalRecords.Value);
-                }
-
-                DumbAction.RehearsalRecords.Value =
-                    ImmutableList<(Address, string)>.Empty;
-                IAccountStateDelta delta = actionEvaluator.EvaluateTxResult(
-                    tx,
-                    default,
-                    1,
-                    new AccountStateDeltaImpl(
-                        ActionEvaluator<DumbAction>.NullAccountStateGetter,
-                        ActionEvaluator<DumbAction>.NullAccountBalanceGetter,
-                        tx.Signer),
-                    addresses[0],
-                    rehearsal: rehearsal);
-                Assert.Equal(
-                    evaluations[3].OutputStates.GetUpdatedStates(),
-                    delta.GetUpdatedStates());
-
-                if (rehearsal)
-                {
-                    Assert.Contains(
-                        (addresses[2], "R"),
-                        DumbAction.RehearsalRecords.Value);
-                }
-                else
-                {
-                    Assert.DoesNotContain(
-                        (addresses[2], "R"),
-                        DumbAction.RehearsalRecords.Value);
-                }
-            }
-        }
-
-        [Fact]
-        public void EvaluateActionsThrowingException()
-        {
-            var action = new ThrowException { ThrowOnRehearsal = false, ThrowOnExecution = true };
-            Transaction<ThrowException> tx = Transaction<ThrowException>.Create(
-                0,
-                _fx.PrivateKey1,
-                null,
-                new[] { action },
-                ImmutableHashSet<Address>.Empty,
-                DateTimeOffset.UtcNow
-            );
-            var hash = new BlockHash(GetRandomBytes(32));
-            var actionEvaluator = new ActionEvaluator<ThrowException>(
-                policyBlockAction: null,
-                stateGetter: ActionEvaluator<ThrowException>.NullStateGetter,
-                balanceGetter: ActionEvaluator<ThrowException>.NullBalanceGetter,
-                trieGetter: null);
-            var nextStates = actionEvaluator.EvaluateTxResult(
-                tx,
-                preEvaluationHash: hash,
-                blockIndex: 123,
-                previousStates: new AccountStateDeltaImpl(
-                    ActionEvaluator<ThrowException>.NullAccountStateGetter,
-                    ActionEvaluator<ThrowException>.NullAccountBalanceGetter,
-                    tx.Signer),
-                miner: GenesisMinerAddress,
-                rehearsal: false);
-
-            Assert.Empty(nextStates.GetUpdatedStates());
         }
 
         [Fact]
